Move global nav related style into its component
