+++ app/scripts/components/common/nav-wrapper.js
@@ -4,7 +4,6 @@ import styled from 'styled-components';
 import PageHeader from './page-header';
 
 import PageLocalNav from '$components/common/page-local-nav';
-import useScrollDirection from '$utils/use-scroll-direction';
 
 const NavWrapper = styled.div`
   position: sticky;
@@ -12,26 +11,12 @@ const NavWrapper = styled.div`
   z-index: 1000;
 `;
 
-const AnimatedPageHeader = styled(PageHeader)`
-  transition: margin-top 0.32s ease-out;
-  margin-top: 0;
-  &.up {
-    margin-top: 0;
-  }
-  &.down {
-    margin-top: calc(
-      -2.5rem - (1rem * var(--base-space-multiplier, 1) * 0.75 * 2)
-    );
-  }
-`;
-
 function PageNavWrapper({ localNavProps }) {
   const renderLocalNav = !!localNavProps;
-  const scrollDir = useScrollDirection();
 
   return (
     <NavWrapper>
-      <AnimatedPageHeader className={scrollDir} />
+      <PageHeader />
       {renderLocalNav && <PageLocalNav {...localNavProps} />}
     </NavWrapper>
   );
